move classes to their right places
