+++ workbenchfx-demo/src/main/java/com/dlsc/workbenchfx/demo/CustomDemo.java
@@ -7,10 +7,10 @@ import com.dlsc.workbenchfx.demo.controls.CustomPage;
 import com.dlsc.workbenchfx.demo.controls.CustomTab;
 import com.dlsc.workbenchfx.demo.controls.CustomTile;
 import com.dlsc.workbenchfx.demo.modules.calendar.CalendarModule;
-import com.dlsc.workbenchfx.modules.gantt.GanttModule;
+import com.dlsc.workbenchfx.demo.modules.gantt.GanttModule;
 import com.dlsc.workbenchfx.demo.modules.helloworld.HelloWorldModule;
 import com.dlsc.workbenchfx.demo.modules.maps.MapsModule;
-import com.dlsc.workbenchfx.modules.patient.PatientModule;
+import com.dlsc.workbenchfx.demo.modules.patient.PatientModule;
 import com.dlsc.workbenchfx.demo.modules.preferences.Preferences;
 import com.dlsc.workbenchfx.demo.modules.preferences.PreferencesModule;
 import com.dlsc.workbenchfx.demo.modules.test.DialogTestModule;
